Utilise the new winit ApplicationHandler API for handling the main loop.
This also moves window creation to `new_events`, because only with an
`ActiveEventLoop` can we create a Window (and therefore... an Engine).

In practice this is not a big deal, because we just get a mutable ref to
the Engine at the beginning of each callback in the MainLoop type.
